% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialize.R
\name{spatialize}
\alias{spatialize}
\title{Query the API and spatialize the data}
\usage{
spatialize(isodatetime, sensor, token_env_var)
}
\arguments{
\item{isodatetime}{A character specifying the isodatetime ("YYYY-MM-DDTHH:MM:SS") set of records you want to spatialize}

\item{sensor}{A character specifying the sensor you want to spatiliaze (tsa, hra)}

\item{token_env_var}{A character specifying the name of the environment var used to store the agromet API token in the ~/.Renviron file}
}
\value{
A dataframe containing the spatialized data.
}
\description{
Query the API and spatialize the data
}
\author{
Thomas Goossens
}
