% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialize.R
\name{spatialize}
\alias{spatialize}
\title{Query the API and spatialize the data}
\usage{
spatialize(isodatetime, sensor)
}
\arguments{
\item{isodatetime}{A character specifying the isodatetime ("YYYY-MM-DDTHH:MM:SS") set of records you want to spatialize}

\item{sensor}{A character specifying the sensor you want to spatiliaze (tsa, hra)}
}
\value{
A dataframe containing the spatialized data.
}
\description{
Query the API and spatialize the data
}
\author{
Thomas Goossens
}
